requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.5
django-cors-headers>=4.3.1
django-environ>=0.11.2
django-celery-results>=2.5.1
django-celery-beat>=2.5.0
celery>=5.3.6
redis>=5.0.1
psycopg2-binary>=2.9.9
drf-nested-routers>=0.94.1